#124 Написал тесты для citiesRouter

+++ backend/routers/cities.ts
@@ -5,13 +5,20 @@ import permit from '../middleware/permit';
 import mongoose from 'mongoose';
 import Location from '../models/Location';
 import Street from '../models/Street';
+import Region from '../models/Region';
 
 const citiesRouter = express.Router();
 
 citiesRouter.get('/', auth, async (req, res, next) => {
+  const areaId = req.query.areaId as string;
+
   try {
-    if (req.query.areaId !== undefined) {
-      const cities = await City.find({ area: req.query.areaId });
+    if (areaId !== undefined) {
+      if (!mongoose.isValidObjectId(areaId)) {
+        return res.status(422).send({ error: 'Некорректный id области.' });
+      }
+
+      const cities = await City.find({ area: areaId });
       return res.send(cities);
     } else {
       const cities = await City.find();
@@ -24,15 +31,15 @@ citiesRouter.get('/', auth, async (req, res, next) => {
 
 citiesRouter.post('/', auth, permit('admin'), async (req, res, next) => {
   try {
-    const cityData = await City.create({
+    const city = await City.create({
       area: req.body.area,
       name: req.body.name,
     });
 
-    return res.send(cityData);
+    return res.status(201).send({ message: 'Новый город успешно создан!', city });
   } catch (e) {
     if (e instanceof mongoose.Error.ValidationError) {
-      return res.status(400).send(e);
+      return res.status(422).send(e);
     } else {
       return next(e);
     }
@@ -40,16 +47,24 @@ citiesRouter.post('/', auth, permit('admin'), async (req, res, next) => {
 });
 
 citiesRouter.delete('/:id', auth, permit('admin'), async (req, res, next) => {
+  const _id = req.params.id as string;
+
+  if (!mongoose.isValidObjectId(_id)) {
+    return res.status(422).send({ error: 'Некорректный id города.' });
+  }
+
   try {
-    const _id = req.params.id as string;
     const city = await City.findOne({ _id });
-    const location = await Location.find({ city: _id });
-    const streets = await Street.find({ city: _id });
+    const location = await Location.findOne({ city: _id });
+    const street = await Street.findOne({ city: _id });
+    const region = await Region.findOne({ city: _id });
+
     if (!city) {
       return res.status(404).send({ error: 'Город не существует в базе.' });
-    } else if (location.length > 0 || streets.length > 0) {
-      return res.status(404).send({ error: 'Город привязан к сущностям ! удаление запрещено' });
+    } else if (location || street || region) {
+      return res.status(409).send({ error: 'Город привязан к другим сущностям! Удаление запрещено.' });
     }
+
     const result = await City.deleteOne({ _id });
     return res.send(result);
   } catch (e) {
